pyproject.toml
[project]
name = "C-d"
version = "0.1.0"
description = "A light weight C_d predictor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.5.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "paddlepaddle>=3.0.0",
    "pytorch-ignite>=0.5.2",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
